[CHORE] : moved room related type to RoomType.tsx & etc
