+++ frontend/type/type.tsx
@@ -53,8 +53,3 @@ export interface ICor {
 export interface IPaddle extends ICor {}
 
 export interface IBall extends ICor {}
-
-export interface ReturnMsgDto {
-  code: number,
-  msg: string,
-}
